added NoteDialog.js, removed dialog component from Note.js and refactored close dialog box listen events to seperate the two types of dialog boxes; moved criminalIDFinder to criminalDataProvider.js
