+++ scripts/criminals/Dialog.js
@@ -1,6 +1,7 @@
 const eventHub = document.querySelector(".container");
+const targetListContainerContentElement = document.querySelector(".listContainer");
 
-// Renders the dialog elements to the DOM for each specific criminal when being called in a for/of loop.
+// Renders the associate alibi dialog elements to the DOM for each specific criminal when being called in a for/of loop.
 export const dialogElement = (criminalObject) => {
     let contentTargetElement = document.querySelector(`#criminal__information--${criminalObject.id}`);
     contentTargetElement.innerHTML +=`
@@ -16,23 +17,6 @@ export const dialogElement = (criminalObject) => {
         </dialog>
         `
 }
-/*
- *  Listens for a "click" event and dispatches the custom event, dialogButtonDetailEvent,
- *  to the eventHub to open a corresponding dialog box.
-*/
-eventHub.addEventListener(
-    "click", 
-    event => {
-    if (event.target.id.startsWith("associates--")) {
-        const [prefix, chosenCriminal] = event.target.id.split("--");
-        const openDialogBox = new CustomEvent("dialogButtonDetailEvent", {
-            detail: {
-                criminal: chosenCriminal
-            }
-        })
-        eventHub.dispatchEvent(openDialogBox);
-    }
-})
 
 // Listens for the custom event, dialogButtonDetailEvent, to open a corresponding dialog box.
 eventHub.addEventListener("dialogButtonDetailEvent", event => {
@@ -40,27 +24,15 @@ eventHub.addEventListener("dialogButtonDetailEvent", event => {
     const theDialog = document.querySelector(dialogSiblingSelector);
     theDialog.showModal();
 })
-/*
-*   Listens for a "click" event and dispatches the custom event, dialogButtonCloseEvent,
-*   to the eventHub to close a corresponding dialog box.
-*/
-eventHub.addEventListener(
+
+// Listens for a "click" event and closes a corresponding dialog box.
+targetListContainerContentElement.addEventListener(
     "click", 
     event => {
     if (event.target.id.startsWith("close-")) {
         const [prefix, chosenDialog] = event.target.id.split("-");
-        const closeDialogBox = new CustomEvent("dialogButtonCloseEvent", {
-            detail: {
-                boxID: chosenDialog
-            }
-        })
-        eventHub.dispatchEvent(closeDialogBox);
+        const theDialogBoxID = `#details--${chosenDialog}`;
+        const theDialogElement = document.querySelector(theDialogBoxID);
+        theDialogElement.close();
     }
-})
-
-// Listens for the custom event, dialogButtonCloseEvent, to close a corresponding dialog box.
-eventHub.addEventListener("dialogButtonCloseEvent", event => {
-    const theDialogBoxID = `#details--${event.detail.boxID}`;
-    const theDialogElement = document.querySelector(theDialogBoxID);
-    theDialogElement.close();
 })
