+++ scripts/criminals/criminalDataProvider.js
@@ -16,3 +16,12 @@ export const getCriminals = () => {
             }
         );
 }
+/*
+ *  A function, findCriminalID, that returns the value from the key/value pair in a criminal object ("criminalId": X,)
+ *  in order to store the criminal's unknown number in the new note added when a user clicks the Save Note button. 
+*/
+export const findCriminalID = (arrayOfCriminalObjects, criminalName) => {
+    //Declair a variable to store the final value, then map over the criminal array in order to find a matching value, then return it.
+    const criminalIdValue = arrayOfCriminalObjects.filter(criminal => criminal.name === criminalName);
+    return criminalIdValue[0].id;
+}
