+++ scripts/notes/Note.js
@@ -6,21 +6,8 @@ export const NoteComponent = noteObject => `
 
         <button id="toggleNote--${noteObject.id}" class="toggleNoteButton">Hide Note</button>
         
-        <button id="editNote--${noteObject.id}" class="editNoteButton">Edit Note</button>
-
-        <dialog class="dialog--note" id="details--${noteObject.id}">
-            <label for="note--date--edit--${noteObject.id}" class="note--date--editLabel">Date:</label>
-            <input type="date" id="note--date--edit--${noteObject.id}" class="note--date--edit" required></br>
-            <label for="note--suspect--edit--${noteObject.id}" class="note--suspect--editLabel">Suspect:</label>
-            <textarea id="note--suspect--edit--${noteObject.id}" class="note--suspect--edit" required>${noteObject.suspect}</textarea></br>
-            <label for="note--text--edit--${noteObject.id}" class="note--text--editLabel">Note:</label></br>
-            <textarea id="note--text--edit--${noteObject.id}" class="note--text--edit" required>${noteObject.noteText}</textarea></br>
-
-        
-            <button id="editNoteSubmit--${noteObject.id}--${noteObject.criminalId}" class="editNoteSubmitButton">Submit Edited Note</button>
-            <button class="button--close button--close--edit" id="close-${noteObject.id}">Close Unedited</button>
-        </dialog>
-
         <button id="deleteNote--${noteObject.id}" class="deleteNoteButton">Delete Note</button>
+
+        <button id="editNote--${noteObject.id}" class="editNoteButton">Edit Note</button>
     </section>
 `
